requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.4.2            # via -r requirements.in
appdirs==1.4.3            # via black
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black
black==19.10b0            # via -r requirements.in
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via -r requirements.in, black, flask, uvicorn
entrypoints==0.3          # via flake8
fastapi==0.54.1           # via -r requirements.in
flake8==3.7.9             # via -r requirements.in
flask==1.1.2              # via -r requirements.in
gevent==1.4.0             # via -r requirements.in
greenlet==0.4.15          # via gevent
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
idna==2.9                 # via requests
isort==4.3.21             # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
kafka-python==2.0.1       # via -r requirements.in
lazy-object-proxy==1.4.3  # via astroid
mako==1.1.2               # via alembic
markupsafe==1.1.1         # via jinja2, mako
mccabe==0.6.1             # via flake8, pylint
pathspec==0.7.0           # via black
prometheus-client==0.7.1  # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
pycodestyle==2.5.0        # via flake8
pydantic==1.5.1           # via fastapi
pyflakes==2.1.1           # via flake8
pylint==2.4.4             # via -r requirements.in
python-dateutil==2.8.1    # via alembic
python-dotenv==0.12.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
regex==2020.2.20          # via black
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via astroid, python-dateutil, structlog
sqlalchemy==1.3.15        # via -r requirements.in, alembic
starlette==0.13.2         # via fastapi
structlog==20.1.0         # via -r requirements.in
toml==0.10.0              # via black
typed-ast==1.4.1          # via astroid, black
urllib3==1.25.8           # via requests
uvicorn==0.11.5           # via -r requirements.in
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
werkzeug==1.0.1           # via flask
wrapt==1.11.2             # via astroid
